add an IterativeAlgorithm class to put in common all the boing stuff of iterative algorithms (for future dev)
