+++ mrfx/__init__.py
@@ -1,3 +1,4 @@
+import mrfx.abstract
 import mrfx.samplers
 import mrfx.models
 import mrfx.estimators
